+++ udata/settings.py
@@ -63,10 +63,14 @@ class Defaults(object):
     CELERY_TASK_DEFAULT_EXCHANGE_TYPE = 'topic'
     CELERY_TASK_DEFAULT_ROUTING_KEY = 'task.default'
     CELERY_TASK_ROUTES = {
-        # High priority for search tasks
+        # High priority for search tasks and mails
         'udata.search.reindex': {
             'queue': 'high',
             'routing_key': 'high.search',
+        },
+        'udata.patch_flask_security.sendmail': {
+            'queue': 'high',
+            'routing_key': 'high.sendmail',
         },
         # Low priority for harvest operations
         'harvest': {